tst_test.sh: Fix IPC init for checkpoints
The introduction of magic at the start of the LTP IPC region did break
the checkpoints for the tst_test.sh shell library. The reason is that
the library creates the IPC backing file with dd and the magic wasn't
present there.

This introduced failures in the pec testcases:

  cn_pec 1 TINFO: Test process events connector
  cn_pec 1 TINFO: Testing fork event (nevents=10)
  tst_test.c:203: TBROK: Invalid shared memory region (bad magic)
  cn_pec 1 TBROK: tst_checkpoint wait 10000 0 failed

This patch fixes the tst_test.sh library so that the IPC region is
created with a correct magic.

Follow-up: bf9589d5bd ("lib: moves test infrastructure states into a shared context structure")
Reported-by: Avinesh Kumar <akumar@suse.de>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>